Implemented basic game logic
Has working basic game logic for player and dealer. Some bugs to work out in terms of timing of different events firing. Also need
to consider refactoring some game logic code to reside in different models. Future goals are also to improve the visuals of the
project with graphics and CSS.